% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyInput.R
\name{bodyInput}
\alias{bodyInput}
\title{An SVG Human Body Input}
\usage{
bodyInput(inputId, data = NULL, ...)
}
\arguments{
\item{inputId}{The input id}

\item{data}{a vector of numbers to pass to the input to color body parts by.}

\item{...}{Passed to [htmltools::div()]

Can be numeric or categorical

  For example

  ```
  bodyInput(
    inputId = "human"
  )
  ```}
}
\value{
The value returned by the input to the Shiny server is either `NULL`
  when no body part is select or a character string containing
  the selected body part
}
\description{
This input operates like a [shiny::radioButtons()] where you can select one of the body parts!
}
