% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-generator.R
\name{bodyPalette}
\alias{bodyPalette}
\title{Color Generator for the Shiny body input}
\usage{
bodyPalette(n, low.col = "#eef4fb", high.col = "#164a85")
}
\arguments{
\item{n}{vector of values to create colors for}

\item{low.col}{the color palette passed from body-input}

\item{high.col}{the color palette passed from body-input}
}
\description{
Color Generator for the Shiny body input
}
