% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\name{human}
\alias{human}
\title{Interactive Human Body Widget}
\usage{
human(
  gender = c("male", "female"),
  shown = c("thyroid_gland", "adrenal_gland"),
  selected = c("adrenal_gland"),
  hovertext = c(thyroid_gland = "Metastases", adrenal_gland = "Primary Tumor"),
  show_color = c(thyroid_gland = "red", adrenal_gland = "grey"),
  select_color = "yellow",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{gender}{One of "male" or "female"}

\item{shown}{The organs that should be shown, e.g. c("brain", "heart").
See ?shinybody_organs for a list of all organs that can be shown.}

\item{selected}{The organs that should be in a selected state.}

\item{hovertext}{Optional. A named vector where the names are shown organs
and the values are the hovertext to show when the organ is hovered.}

\item{show_color}{Either a single color (e.g. "black") or a named vactor of
colors, where the names are shown organs and the values are colors (e.g.
c("brain" = "green", "heart" = "red"))}

\item{select_color}{The color that should be applied to organs with the
selected state}

\item{width}{Widget width}

\item{height}{Widget height}

\item{elementId}{ID of the widget}
}
\description{
This widget visualizes an SVG-based human body, highlights specific body parts,
and displays associated participant data.
}
