% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patients}
\alias{patients}
\title{Example data set of patients}
\format{
\subsection{\code{patients}}{

A data frame with 16 rows and 5 columns:
\describe{
\item{patient_id }{A unique patient identifier}
\item{gender}{"male" or "female"}
\item{age}{Patient age}
\item{height}{Patient height in inches}
\item{weight}{Patient weight in lbs}
}
}
}
\usage{
patients
}
\description{
A randomly generated dataset of patient details
}
\keyword{datasets}
