% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shinybody_organs}
\alias{shinybody_organs}
\title{Organs available in shinybody}
\format{
\subsection{\code{shinybody_organs}}{

A list with two elements: \code{male} and \code{female}.
\describe{
\item{male}{A list of the organs that can be shown on the male body avatar}
\item{female}{A list of the organs that can be shown on the male body avatar}
}
}
}
\usage{
shinybody_organs
}
\description{
A list of the organs that shinybody can display
}
\keyword{datasets}
