% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumors}
\alias{tumors}
\title{Example data set of tumors}
\format{
\subsection{\code{tumors}}{

A data frame with 39 rows and 5 columns:
\describe{
\item{patient_id }{A unique patient identifier}
\item{tumor_id}{A unique tumor identifier}
\item{tumor_location}{The organ affected by the tumor}
\item{is_primary_tumor}{TRUE if the tumor is the patient's primary cancer site, otherwise FALSE}
\item{stage}{The stage of the tumor (I, II, or III)}
}
}
}
\usage{
tumors
}
\description{
A randomly generated dataset of tumors to use in examples
}
\keyword{datasets}
