% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyInput.R
\name{bodyInput}
\alias{bodyInput}
\title{An SVG Human Body Input}
\usage{
bodyInput(inputId, data = NULL, ...)
}
\arguments{
\item{inputId}{The input id}

\item{data}{a vector of 13 numbers to pass to the input to color body parts by:

 \enumerate{
  \item head
  \item left-shoulder
  \item right-shoulder
  \item left-arm
  \item right-arm
  \item left-hand
  \item right-hand
  \item chest
  \item stomach
  \item left-leg
  \item right-lef
  \item left-foot
  \item right-foot
}}

\item{...}{Passed to \code{htmltools::div()}}
}
\value{
The value returned by the input to the Shiny server is either `NULL`
  when no body part is select or a character string containing
  the selected body part
}
\description{
This input operates like a \code{shiny::radioButtons()} where you can select one of the body parts!
}
\examples{
\dontrun{
ui <- function() {
 fluidPage(
   bodyInput("human", data = c(10,20,40,60,90,100,25,50,15,20,70,70,30))),
   verbatimTextOutput("debug")
 )
}

server <- function(input, output) {
 output$debug <- renderText(input$human)
}
shinyApp(ui = ui, server = server)
}

}
