% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-generator.R
\name{bodyPalette}
\alias{bodyPalette}
\title{Color Generator for the Shiny body input}
\usage{
bodyPalette(n, name = c("body.colors"), pal = "Blues")
}
\arguments{
\item{n}{vector of values to create colors for}

\item{name}{the name of the color function}

\item{pal}{the color palette passed from body-input}
}
\description{
Color Generator for the Shiny body input
}
